+++ src/main/java/fr/alasdiablo/janoeo/foundation/Foundation.java
@@ -10,22 +10,22 @@ import fr.alasdiablo.janoeo.foundation.init.FoundationBlocks;
 import fr.alasdiablo.janoeo.foundation.init.FoundationItems;
 import fr.alasdiablo.janoeo.foundation.world.WorldGen;
 import fr.alasdiablo.janoeo.foundation.world.gen.FoundationFeatures;
-import mcp.MethodsReturnNonnullByDefault;
-import net.minecraft.block.Block;
-import net.minecraft.block.Blocks;
+import net.minecraft.MethodsReturnNonnullByDefault;
 import net.minecraft.data.DataGenerator;
-import net.minecraft.item.Item;
-import net.minecraft.item.ItemGroup;
-import net.minecraft.item.ItemStack;
-import net.minecraft.item.Items;
+import net.minecraft.world.item.CreativeModeTab;
+import net.minecraft.world.item.Item;
+import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.item.Items;
+import net.minecraft.world.level.block.Block;
+import net.minecraft.world.level.block.Blocks;
 import net.minecraftforge.common.data.ExistingFileHelper;
 import net.minecraftforge.event.RegistryEvent;
 import net.minecraftforge.eventbus.api.IEventBus;
 import net.minecraftforge.fml.ModList;
 import net.minecraftforge.fml.common.Mod;
 import net.minecraftforge.fml.event.lifecycle.FMLCommonSetupEvent;
-import net.minecraftforge.fml.event.lifecycle.GatherDataEvent;
 import net.minecraftforge.fml.javafmlmod.FMLJavaModLoadingContext;
+import net.minecraftforge.forge.event.lifecycle.GatherDataEvent;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
@@ -37,14 +37,14 @@ public class Foundation {
 
     public static final Logger logger = LogManager.getLogger(Registries.MOD_ID);
 
-    public static final ItemGroup ITEMS_GROUP = new GroundItemGroup("janoeo.foundation.item") {
+    public static final CreativeModeTab ITEMS_GROUP = new GroundItemGroup("janoeo.foundation.item") {
         @Override
         public ItemStack makeIcon() {
             return new ItemStack(Items.COAL);
         }
     };
 
-    public static final ItemGroup BLOCKS_GROUP = new GroundItemGroup("janoeo.foundation.block") {
+    public static final CreativeModeTab BLOCKS_GROUP = new GroundItemGroup("janoeo.foundation.block") {
         @Override
         public ItemStack makeIcon() {
             return new ItemStack(Blocks.COAL_BLOCK);
@@ -104,7 +104,7 @@ public class Foundation {
 
     private void setup(final FMLCommonSetupEvent commonSetupEvent) {
         WorldGen.init();
-        if (Compat.JUST_ENOUGH_RESOURCES) JERCompat.init();
+        // if (Compat.JUST_ENOUGH_RESOURCES) JERCompat.init();
     }
 
     private void initFeatures(RegistryEvent.NewRegistry newRegistry) {