+++ src/main/java/fr/alasdiablo/janoeo/foundation/data/FoundationBlockModelProvider.java
@@ -3,7 +3,7 @@ package fr.alasdiablo.janoeo.foundation.data;
 import static fr.alasdiablo.janoeo.foundation.Registries.*;
 
 import net.minecraft.data.DataGenerator;
-import net.minecraft.util.ResourceLocation;
+import net.minecraft.resources.ResourceLocation;
 import net.minecraftforge.client.model.generators.BlockModelProvider;
 import net.minecraftforge.common.data.ExistingFileHelper;