+++ src/main/java/fr/alasdiablo/janoeo/foundation/data/language/EnglishProvider.java
@@ -3,7 +3,7 @@ package fr.alasdiablo.janoeo.foundation.data.language;
 import fr.alasdiablo.janoeo.foundation.Foundation;
 import fr.alasdiablo.janoeo.foundation.Registries;
 import net.minecraft.data.DataGenerator;
-import net.minecraft.util.text.TranslationTextComponent;
+import net.minecraft.network.chat.TranslatableComponent;
 import net.minecraftforge.common.data.LanguageProvider;
 
 import static fr.alasdiablo.janoeo.foundation.init.FoundationItems.*;
@@ -76,7 +76,7 @@ public class EnglishProvider extends LanguageProvider {
         this.add(TINY_REDSTONE_ORE, "Tiny redstone ore");
 
 
-        this.add(((TranslationTextComponent) Foundation.ITEMS_GROUP.getDisplayName()).getKey(), "Janoeo Foundation - Items");
-        this.add(((TranslationTextComponent) Foundation.BLOCKS_GROUP.getDisplayName()).getKey(), "Janoeo Foundation - Blocks");
+        this.add(((TranslatableComponent) Foundation.ITEMS_GROUP.getDisplayName()).getKey(), "Janoeo Foundation - Items");
+        this.add(((TranslatableComponent) Foundation.BLOCKS_GROUP.getDisplayName()).getKey(), "Janoeo Foundation - Blocks");
     }
 }